Refactor rendering entity types and selection management
- Removed the obsolete pick_entity_type.hpp and its implementation.
- Introduced render_types.hpp to define RenderEntityType and RenderEntityTypeMask enums, consolidating entity type definitions for geometry and mesh.
- Updated RenderSelectManager to utilize the new RenderEntityType and RenderEntityTypeMask, including changes to selection methods and type checks.
- Modified SelectManagerService to accommodate changes in entity type handling, ensuring compatibility with the new RenderEntityType.
- Adjusted geometry and mesh type handling to support uint64_t for IDs and UIDs, enhancing type safety and consistency across the application.
- Updated mesh element type handling to replace invalid types with None, improving error handling and clarity.
- Ensured all relevant methods and functions reflect the new entity type structure, including conversion functions and selection logic.

+++ include/geometry/geometry_types.hpp
@@ -66,7 +66,7 @@ enum class EntityType : uint8_t {
  * EntityId provides a globally unique identifier across all entity types.
  * It can be used to quickly locate any entity in the geometry system.
  */
-using EntityId = uint32_t;
+using EntityId = uint64_t;
 
 /**
  * @brief Type-scoped unique identifier within the same entity type
@@ -75,7 +75,7 @@ using EntityId = uint32_t;
  * and edge UID 1 are different entities. Combined with EntityType, it forms
  * a complete entity reference.
  */
-using EntityUID = uint32_t;
+using EntityUID = uint64_t;
 
 /// Invalid/null EntityId constant
 constexpr EntityId INVALID_ENTITY_ID = 0;
@@ -101,7 +101,7 @@ constexpr EntityUID INVALID_ENTITY_UID = 0;
  * @param type The entity type to query
  * @return The maximum assigned EntityUID for the given type
  */
-[[nodiscard]] uint32_t getMaxIdByType(EntityType type);
+[[nodiscard]] uint64_t getMaxIdByType(EntityType type);
 
 /**
  * @brief Reset UID generator for a specific type (for testing purposes)
